// 25032026/sce-front-auditoria/sce-front-auditoria/src/app/model/control-calidad/DataPopupVerResolucion.ts
export interface DataPopupVerResolucion {
    idActa: number;
    idResolucion: number;
    numeroActa: string;
    nombreResolucion: string;
    eleccion: string;
    numeroElectores: number;
    numeroElectoresAusentes: number;
    ubigeo: string;
    archivoResolucion: any;
    numeroExpediente: string;
}